Extract geometry functions from render-net to lib/geometry
Moves extractShapes, extractPackages, extractViaPadSizes,
extractComponents, and transformPin to a shared geometry module.
Prepares for reuse by query-via-in-pad and pad geometry features.
