+++ src/tools/render-net.ts
@@ -17,6 +17,14 @@ import { Resvg, initWasm } from "@resvg/resvg-wasm";
 import { z } from "zod";
 import type { ErrorResult, RenderNetResult } from "./lib/types.js";
 import { isErrorResult } from "./lib/types.js";
+import type { Point, Shape, PinDef, ComponentPlacement } from "./lib/geometry.js";
+import {
+  extractShapes,
+  extractPackages,
+  extractViaPadSizes,
+  extractComponents,
+  transformPin,
+} from "./lib/geometry.js";
 import { attr, numAttr, loadAllLines, scanLines } from "./lib/xml-utils.js";
 import {
   extractMicronFactorFromLines,
@@ -117,33 +125,10 @@ const buildLayerColors = (layers: string[]): Map<string, string> => {
 // Types
 // =============================================================================
 
-interface Point {
-  x: number;
-  y: number;
-}
 interface ArcPoint extends Point {
   centerX: number;
   centerY: number;
   clockwise: boolean;
-}
-interface Shape {
-  type: "rect" | "circle";
-  width: number;
-  height: number;
-}
-interface PinDef {
-  offsetX: number;
-  offsetY: number;
-  shapeId: string;
-}
-interface ComponentPlacement {
-  refdes: string;
-  packageRef: string;
-  x: number;
-  y: number;
-  rotation: number;
-  mirror: boolean;
-  layer: string;
 }
 interface Trace {
   layer: string;
@@ -160,200 +145,6 @@ interface Via {
 // =============================================================================
 // Extraction passes (in-memory line scanning)
 // =============================================================================
-
-const extractShapes = (lines: string[], f: number): Map<string, Shape> => {
-  const shapes = new Map<string, Shape>();
-  let currentId = "";
-
-  scanLines(lines, (line) => {
-    if (line.includes("<EntryStandard ")) {
-      currentId = attr(line, "id") ?? "";
-    }
-    if (currentId && line.includes("<RectCenter ")) {
-      const w = numAttr(line, "width");
-      const h = numAttr(line, "height");
-      if (w !== undefined && h !== undefined) {
-        shapes.set(currentId, { type: "rect", width: w * f, height: h * f });
-      }
-      currentId = "";
-    }
-    if (currentId && line.includes("<Circle ")) {
-      const d = numAttr(line, "diameter");
-      if (d !== undefined) {
-        shapes.set(currentId, { type: "circle", width: d * f, height: d * f });
-      }
-      currentId = "";
-    }
-    if (line.includes("</Content>")) return false;
-  });
-
-  return shapes;
-};
-
-const extractPackages = (lines: string[]): Map<string, Map<string, PinDef>> => {
-  const packages = new Map<string, Map<string, PinDef>>();
-  let currentPkg = "";
-  let inPad = false;
-  let inPin = false;
-  let padPin = "";
-  let padOffset: Point = { x: 0, y: 0 };
-  let padShapeId = "";
-
-  const commitPad = () => {
-    if (currentPkg && padPin && padShapeId) {
-      packages.get(currentPkg)!.set(padPin, {
-        offsetX: padOffset.x,
-        offsetY: padOffset.y,
-        shapeId: padShapeId,
-      });
-    }
-    padPin = "";
-    padOffset = { x: 0, y: 0 };
-    padShapeId = "";
-  };
-
-  scanLines(lines, (line) => {
-    if (line.includes("<Package ")) {
-      currentPkg = attr(line, "name") ?? "";
-      if (currentPkg && !packages.has(currentPkg)) {
-        packages.set(currentPkg, new Map());
-      }
-    }
-    if (line.includes("</Package>")) {
-      currentPkg = "";
-    }
-
-    if (!currentPkg) return;
-
-    if (line.includes("<Pad") && (line.includes("<Pad>") || line.includes("<Pad "))) {
-      inPad = true;
-      padPin = "";
-      padOffset = { x: 0, y: 0 };
-      padShapeId = "";
-    }
-    if (inPad) {
-      if (line.includes("<PinRef ")) {
-        padPin = attr(line, "pin") ?? "";
-      }
-      if (line.includes("<Location ")) {
-        const x = numAttr(line, "x");
-        const y = numAttr(line, "y");
-        if (x !== undefined) padOffset.x = x;
-        if (y !== undefined) padOffset.y = y;
-      }
-      if (line.includes("<StandardPrimitiveRef ")) {
-        padShapeId = attr(line, "id") ?? "";
-      }
-      if (line.includes("</Pad>")) {
-        commitPad();
-        inPad = false;
-      }
-    }
-
-    if (!inPad && line.includes("<Pin ") && line.includes("number=")) {
-      inPin = true;
-      padPin = attr(line, "number") ?? "";
-      padOffset = { x: 0, y: 0 };
-      padShapeId = "";
-    }
-    if (inPin) {
-      if (line.includes("<Location ")) {
-        const x = numAttr(line, "x");
-        const y = numAttr(line, "y");
-        if (x !== undefined) padOffset.x = x;
-        if (y !== undefined) padOffset.y = y;
-      }
-      if (line.includes("<StandardPrimitiveRef ")) {
-        padShapeId = attr(line, "id") ?? "";
-      }
-      if (line.includes("</Pin>")) {
-        commitPad();
-        inPin = false;
-      }
-    }
-
-    if (line.includes("<Component ") && line.includes("refDes=")) return false;
-  });
-
-  return packages;
-};
-
-const extractViaPadSizes = (
-  lines: string[]
-): Map<string, { padShapeId: string; drillDiameter: number }> => {
-  const viaPads = new Map<string, { padShapeId: string; drillDiameter: number }>();
-  let currentName = "";
-  let currentDrill = 0;
-  let foundRegular = false;
-
-  scanLines(lines, (line) => {
-    if (line.includes("<PadStackDef ")) {
-      currentName = attr(line, "name") ?? "";
-      currentDrill = 0;
-      foundRegular = false;
-    }
-    if (currentName && line.includes("<PadstackHoleDef ")) {
-      currentDrill = numAttr(line, "diameter") ?? 0;
-    }
-    if (currentName && !foundRegular && line.includes('padUse="REGULAR"')) {
-      foundRegular = true;
-    }
-    if (currentName && foundRegular && line.includes("<StandardPrimitiveRef ")) {
-      const id = attr(line, "id") ?? "";
-      if (id && currentDrill > 0) {
-        viaPads.set(currentName, { padShapeId: id, drillDiameter: currentDrill });
-      }
-      foundRegular = false;
-    }
-    if (line.includes("</PadStackDef>")) {
-      currentName = "";
-    }
-    if (line.includes("<Package ")) return false;
-  });
-
-  return viaPads;
-};
-
-const extractComponents = (lines: string[], f: number): ComponentPlacement[] => {
-  const components: ComponentPlacement[] = [];
-  let current: ComponentPlacement | null = null;
-
-  scanLines(lines, (line) => {
-    if (line.includes("<Component ") && line.includes("refDes=")) {
-      const refdes = attr(line, "refDes");
-      if (refdes) {
-        current = {
-          refdes,
-          packageRef: attr(line, "packageRef") ?? "",
-          x: 0,
-          y: 0,
-          rotation: 0,
-          mirror: false,
-          layer: attr(line, "layerRef") ?? "",
-        };
-      }
-    }
-    if (current) {
-      if (line.includes("<Location ")) {
-        const x = numAttr(line, "x");
-        const y = numAttr(line, "y");
-        if (x !== undefined) current.x = x * f;
-        if (y !== undefined) current.y = y * f;
-      }
-      if (line.includes("<Xform ")) {
-        current.rotation = numAttr(line, "rotation") ?? 0;
-        current.mirror = attr(line, "mirror") === "true";
-      }
-      if (line.includes("</Component>")) {
-        components.push(current);
-        current = null;
-      }
-    }
-    if (line.includes("<PhyNetGroup ") || line.includes("</Step>")) return false;
-  });
-
-  return components;
-};
 
 const extractProfile = (
   lines: string[],
@@ -509,19 +300,6 @@ const extractNetPins = (lines: string[], netName: string): { refdes: string; pin
 // =============================================================================
 // Geometry helpers
 // =============================================================================
-
-const transformPin = (comp: ComponentPlacement, pinDef: PinDef, f: number): Point => {
-  const rad = (comp.rotation * Math.PI) / 180;
-  const cos = Math.cos(rad);
-  const sin = Math.sin(rad);
-  let dx = pinDef.offsetX * f;
-  const dy = pinDef.offsetY * f;
-  if (comp.mirror) dx = -dx;
-  return {
-    x: comp.x + dx * cos - dy * sin,
-    y: comp.y + dx * sin + dy * cos,
-  };
-};
 
 const svgArc = (prev: Point, arc: ArcPoint): string => {
   const r = Math.sqrt((arc.centerX - prev.x) ** 2 + (arc.centerY - prev.y) ** 2);
